#385 (New Frontend) Repository Activity: - moved new vis to correct folder - added message header to commits, to create similar kind of data from pouch as BE - minor adjustments for merge
