+++ binocular-frontend-new/src/plugins/visualizationPlugins/issues/burndown/assets/thumbnail.ai
