+++ binocular-frontend-new/src/plugins/visualizationPlugins/issues/burndown/src/chart/components/BurndownChartDataPoint.tsx
