+++ binocular-frontend-new/src/plugins/visualizationPlugins/issues/burndown/src/chart/helper/groupIssuesByGranularity.ts
