+++ binocular-frontend-new/src/plugins/visualizationPlugins/issues/burndown/src/chart/types.ts
@@ -1,5 +1,5 @@
 import type { Moment } from 'moment';
-import type { DataPluginIssue } from '../../../../interfaces/dataPluginInterfaces/dataPluginIssues';
+import type { DataPluginIssue } from '../../../../../interfaces/dataPluginInterfaces/dataPluginIssues';
 
 export interface MappedIssue extends Omit<DataPluginIssue, 'createdAt' | 'closedAt'> {
   createdAt: Moment;
