+++ binocular-frontend-new/src/plugins/visualizationPlugins/issues/burndown/src/help/help.tsx
