+++ binocular-frontend-new/src/plugins/visualizationPlugins/issues/burndown/src/reducer/index.ts
@@ -1,6 +1,5 @@
 import { createSlice, type PayloadAction } from '@reduxjs/toolkit';
-import type { DataPluginIssue } from '../../../../interfaces/dataPluginInterfaces/dataPluginIssues.ts';
-import type { DataPluginMergeRequest } from '../../../../interfaces/dataPluginInterfaces/dataPluginMergeRequests.ts';
+import type { DataPluginIssue } from '../../../../../interfaces/dataPluginInterfaces/dataPluginIssues.ts';
 
 export enum DataState {
   EMPTY,
@@ -15,25 +14,22 @@ interface DateRange {
 
 export interface IssuesState {
   issues: DataPluginIssue[];
-  mergeRequests: DataPluginMergeRequest[];
   dateRange: DateRange;
   dataState: DataState;
 }
 
 const initialState: IssuesState = {
   issues: [],
-  mergeRequests: [],
   dateRange: { from: new Date().toISOString(), to: new Date().toISOString() },
   dataState: DataState.EMPTY,
 };
 
-export const issuesSlice = createSlice({
-  name: 'sprints',
+export const burndownSlice = createSlice({
+  name: 'burndown',
   initialState,
   reducers: {
-    setIssues: (state, { payload: { issues, mergeRequests } }: PayloadAction<Pick<IssuesState, 'issues' | 'mergeRequests'>>) => {
+    setIssues: (state, { payload: issues }: PayloadAction<IssuesState['issues']>) => {
       state.issues = issues;
-      state.mergeRequests = mergeRequests;
     },
     setDateRange: (state, action: PayloadAction<DateRange>) => {
       state.dateRange = action.payload;
@@ -44,5 +40,5 @@ export const issuesSlice = createSlice({
   },
 });
 
-export const { setIssues, setDateRange, setDataState } = issuesSlice.actions;
-export default issuesSlice.reducer;
+export const { setIssues, setDateRange, setDataState } = burndownSlice.actions;
+export default burndownSlice.reducer;
