+++ binocular-frontend-new/src/plugins/visualizationPlugins/issues/burndown/src/settings/settings.tsx
