+++ binocular-frontend-new/src/plugins/visualizationPlugins/issues/sprints/assets/thumbnail.ai
