+++ binocular-frontend-new/src/plugins/visualizationPlugins/issues/sprints/src/chart/Chart.tsx
@@ -1,16 +1,17 @@
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { DataState, type IssuesState } from '../reducer';
-import type { BurndownSettings } from '../settings/settings.tsx';
-import { handelPopoutResizing } from '../../../../utils/resizing.ts';
-import type { VisualizationPluginProperties } from '../../../../interfaces/visualizationPluginInterfaces/visualizationPluginProperties.ts';
-import type { DataPluginIssue } from '../../../../interfaces/dataPluginInterfaces/dataPluginIssues.ts';
-import { getDataSlice } from '../../../simpleVisualizationPlugin/src/reducer/index.ts';
-import moment, { type unitOfTime } from 'moment';
+import type { SprintSettings } from '../settings/settings.tsx';
+import { handelPopoutResizing } from '../../../../../utils/resizing.ts';
+import type { VisualizationPluginProperties } from '../../../../../interfaces/visualizationPluginInterfaces/visualizationPluginProperties.ts';
+import { SprintChart } from './SprintChart.tsx';
+import type { DataPluginIssue } from '../../../../../interfaces/dataPluginInterfaces/dataPluginIssues.ts';
+import { getDataSlice } from '../../../../simpleVisualizationPlugin/src/reducer';
+import moment from 'moment';
 import * as React from 'react';
-import { BurndownChart } from './BurndownChart.tsx';
+import { groupSimilarLabels } from './helper/groupSimilarLabels.ts';
 
-const Chart = (props: VisualizationPluginProperties<BurndownSettings, DataPluginIssue>) => {
+const Chart = (props: VisualizationPluginProperties<SprintSettings, DataPluginIssue>) => {
   /*
    * Creating Dispatch and Root State for interaction with the reducer State
    */
@@ -27,6 +28,7 @@ const Chart = (props: VisualizationPluginProperties<BurndownSettings, DataPlugin
    */
   //Redux Global State
   const issues = useSelector<RootState, IssuesState['issues']>((data) => data.plugin.issues);
+  const mergeRequests = useSelector<RootState, IssuesState['mergeRequests']>((data) => data.plugin.mergeRequests);
   const dataState = useSelector<RootState, IssuesState['dataState']>((state: RootState) => state.plugin.dataState);
 
   /**
@@ -61,6 +63,12 @@ const Chart = (props: VisualizationPluginProperties<BurndownSettings, DataPlugin
     dispatch({ type: 'REFRESH' });
   }, [props.dataConnection]);
 
+  const groupedLabels = groupSimilarLabels(
+    [...new Set(issues.flatMap((i) => i.labels))],
+    props.settings.maxNumberOfDifferencesBetweenLabels,
+    props.settings.minNumberOfLabelsPerGroup,
+  );
+
   return (
     <>
       <div className={'w-full h-full flex justify-center items-center'} ref={props.chartContainerRef}>
@@ -71,16 +79,19 @@ const Chart = (props: VisualizationPluginProperties<BurndownSettings, DataPlugin
           </div>
         )}
         {dataState === DataState.COMPLETE &&
-          (issues.length > 0 ? (
-            <BurndownChart
+          (issues.length > 0 || mergeRequests.length > 0 ? (
+            <SprintChart
+              authors={props.authorList}
               issues={issues}
+              mergeRequests={mergeRequests}
+              coloringMode={props.settings.coloringMode}
               sprints={props.sprintList}
               fromDate={moment(props.parameters.parametersDateRange.from)}
               toDate={moment(props.parameters.parametersDateRange.to)}
               showSprints={props.settings.showSprints}
               width={chartWidth}
               height={chartHeight}
-              granularity={props.parameters.parametersGeneral.granularity as unitOfTime.Base}
+              groupedLabels={groupedLabels}
             />
           ) : (
             <div>No Data matching the selected Parameters!</div>
