+++ binocular-frontend-new/src/plugins/visualizationPlugins/issues/sprints/src/chart/components/SprintChartMergeRequest.tsx
@@ -2,7 +2,7 @@ import * as d3 from 'd3';
 import * as React from 'react';
 import { margin } from '../SprintChart';
 import type { MappedDataPluginMergeRequest } from '../types';
-import type { AuthorType } from '../../../../../../types/data/authorType';
+import type { AuthorType } from '../../../../../../../types/data/authorType';
 import classes from './sprintChartMergeRequest.module.css';
 
 export const SprintChartMergeRequest: React.FC<{
