+++ binocular-frontend-new/src/plugins/visualizationPlugins/issues/sprints/src/chart/components/sprintChartIssue.module.css
