+++ binocular-frontend-new/src/plugins/visualizationPlugins/issues/sprints/src/chart/helper/aggregateTimeTrackingData.ts
@@ -1,4 +1,4 @@
-import type { TimeTrackingData } from '../../../../authorBehaviour/timeSpent/src/utilities/dataConverter';
+import type { TimeTrackingData } from '../../../../../authorBehaviour/timeSpent/src/utilities/dataConverter';
 
 export const aggregateTimeTrackingData = (timeTrackingData: TimeTrackingData[]) => {
   const aggregatedTimeTrackingData = new Map<string, number>();
