+++ binocular-frontend-new/src/plugins/visualizationPlugins/issues/sprints/src/chart/helper/groupSimilarLabels.ts
