+++ binocular-frontend-new/src/plugins/visualizationPlugins/issues/sprints/src/reducer/index.ts
@@ -1,5 +1,6 @@
 import { createSlice, type PayloadAction } from '@reduxjs/toolkit';
-import type { DataPluginIssue } from '../../../../interfaces/dataPluginInterfaces/dataPluginIssues.ts';
+import type { DataPluginIssue } from '../../../../../interfaces/dataPluginInterfaces/dataPluginIssues.ts';
+import type { DataPluginMergeRequest } from '../../../../../interfaces/dataPluginInterfaces/dataPluginMergeRequests.ts';
 
 export enum DataState {
   EMPTY,
@@ -14,22 +15,25 @@ interface DateRange {
 
 export interface IssuesState {
   issues: DataPluginIssue[];
+  mergeRequests: DataPluginMergeRequest[];
   dateRange: DateRange;
   dataState: DataState;
 }
 
 const initialState: IssuesState = {
   issues: [],
+  mergeRequests: [],
   dateRange: { from: new Date().toISOString(), to: new Date().toISOString() },
   dataState: DataState.EMPTY,
 };
 
-export const burndownSlice = createSlice({
-  name: 'burndown',
+export const issuesSlice = createSlice({
+  name: 'sprints',
   initialState,
   reducers: {
-    setIssues: (state, { payload: issues }: PayloadAction<IssuesState['issues']>) => {
+    setIssues: (state, { payload: { issues, mergeRequests } }: PayloadAction<Pick<IssuesState, 'issues' | 'mergeRequests'>>) => {
       state.issues = issues;
+      state.mergeRequests = mergeRequests;
     },
     setDateRange: (state, action: PayloadAction<DateRange>) => {
       state.dateRange = action.payload;
@@ -40,5 +44,5 @@ export const burndownSlice = createSlice({
   },
 });
 
-export const { setIssues, setDateRange, setDataState } = burndownSlice.actions;
-export default burndownSlice.reducer;
+export const { setIssues, setDateRange, setDataState } = issuesSlice.actions;
+export default issuesSlice.reducer;
