+++ binocular-frontend-new/src/plugins/visualizationPlugins/issues/sprints/src/saga/index.ts
@@ -1,8 +1,8 @@
 import { put, takeEvery, fork, call, select } from 'redux-saga/effects';
 import { DataState, type IssuesState, setDataState, setDateRange, setIssues } from '../reducer';
-import type { DataPlugin } from '../../../../interfaces/dataPlugin.ts';
-import type { DataPluginIssue } from '../../../../interfaces/dataPluginInterfaces/dataPluginIssues.ts';
-import type { DataPluginMergeRequest } from '../../../../interfaces/dataPluginInterfaces/dataPluginMergeRequests.ts';
+import type { DataPlugin } from '../../../../../interfaces/dataPlugin.ts';
+import type { DataPluginIssue } from '../../../../../interfaces/dataPluginInterfaces/dataPluginIssues.ts';
+import type { DataPluginMergeRequest } from '../../../../../interfaces/dataPluginInterfaces/dataPluginMergeRequests.ts';
 
 export default function* (dataConnection: DataPlugin) {
   yield fork(() => watchRefresh(dataConnection));
